lesson19_project_hard_source/requirements.txt
aaniso8601==9.0.1
attrs==21.2.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.0.12
click==8.0.3
colorama==0.4.6
cryptography==35.0.0
DateTime==5.0
flake8==4.0.1
Flask==2.0.2
flask-restx==0.5.1
Flask-SQLAlchemy==2.5.1
greenlet==2.0.2
idna==3.4
importlib-metadata==4.8.1
importlib-resources==5.4.0
isort==5.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.1.2
MarkupSafe==2.1.2
marshmallow==3.13.0
mccabe==0.6.1
prettytable==3.6.0
pycodestyle==2.8.0
pycparser==2.21
pyflakes==2.4.0
PyJWT==2.3.0
pyrsistent==0.18.0
pytz==2021.3
requests==2.26.0
six==1.16.0
SQLAlchemy==1.4.25
typing_extensions==4.4.0
urllib3==1.26.14
wcwidth==0.2.6
Werkzeug==2.0.2
zipp==3.6.0
zope.interface==5.5.2